[LT] Add a trie data structure for caching IR nodes
Summary: TrieCache provides a way to look up an IR node before we
actually create it. If the lookup hits in TrieCache, we reuse the
existing node and move the current pointer in TrieCache to point to that
node; if the lookup misses, we create a new node and insert it into TrieCache.

ghstack-source-id: 9b4d3d8285bb805ebb0ea0ceb4a91c1ad8974bbd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/76542